Bug fixes, boot hardening, and ASCII art polish
- Fix rootfs magic: NRFS -> CRFS (rootfs was never mounting)
- Fix PIC mask: 0xF9 -> 0xFC to unmask IRQ0/1 (timer + keyboard)
- Fix grub.cfg: add rootfs cmdline tag, set gfxpayload=text
- Fix kernel.c int-to-pointer casts with (uintptr_t)
- Fix keyboard_init hang: add timeouts to all PS/2 controller waits
- Add keyboard layout selection with auto-timeout to US default
- Add mouse_render() calls during keyboard idle loops
- Add hardware cursor update on console putchar
- Fix vmm.c dead code and unused variables
- Fix 2048 game slide detection bug in games.c
- Replace UTF-8 box-drawing chars with ASCII borders for VGA compat
- Clean up ASCII logos across boot, shell, panic, exception, games
- Add keyboard_scancode_to_char() API

+++ src/app_htop.c
@@ -7,8 +7,8 @@
 #include <stdbool.h>
 #include <stdint.h>
 
-extern unsigned long sys_mem_lower;
-extern unsigned long sys_mem_upper;
+extern uint32_t sys_mem_lower;
+extern uint32_t sys_mem_upper;
 
 #define BAR_WIDTH 30
 #define NUM_TASKS 12
